Move plugin/v2 to daemon/pkg/plugin/v2
