+++ daemon/pkg/plugin/v2/plugin.go
