+++ plugin/store.go
@@ -7,10 +7,10 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/distribution/reference"
+	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/plugins"
-	v2 "github.com/docker/docker/plugin/v2"
 	"github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
 )
